Adding in copy of DB directory with DB.php
This is not code that I have written. I cloned all the files in this
particular commit from: https://github.com/pear/DB but wanted to
include it in this repo to allow for a more effective download.

(If it is not appropriate, please let me know, and I can remove these
files and replace with a pointer).